BigInt: fix, other TypedArray should be Big

+++ test/built-ins/TypedArrayConstructors/ctors-bigint/typedarray-arg/other-ctor-buffer-ctor-access-throws.js
@@ -26,7 +26,7 @@ features: [BigInt, TypedArray]
 ---*/
 
 testWithBigIntTypedArrayConstructors(function(TA) {
-  var OtherCtor = TA === Int8Array ? Int16Array : Int8Array;
+  var OtherCtor = TA === BigInt64Array ? BigUint64Array : BigInt64Array;
   var sample = new OtherCtor();
 
   Object.defineProperty(sample.buffer, "constructor", {
